[ Feat ] refactoring
