+++ Frontend/src/components/app/index.jsx
@@ -3,10 +3,8 @@ import { Routes, Route } from 'react-router-dom';
 import Login from '../../pages/login';
 import SignIn from '../../pages/signin';
 import Dashboard from '../../pages/dashboard';
-import LayoutModal from '../modal-form/index';
-import BasicModal from '../modal-dashboard';
 import ModalDashboardPage from '../../pages/modal-dashboard-page';
-import SuccessPopup from "../success-popup";
+import SuccessPopup from "../popup-success";
 
 function App() {
   return (
@@ -16,7 +14,7 @@ function App() {
         <Route path="register" element={<SignIn />} />
         <Route path="dashboard" element={<Dashboard />} />
         <Route path="modal" element={<ModalDashboardPage />} />
-          <Route path="mini" element={<SuccessPopup />} />
+        <Route path="mini" element={<SuccessPopup />} />
       </Routes>
     </div>
   );
